+++ docs/assets/search.js
@@ -1,0 +1,1 @@
+window.searchData = JSON.parse("{\"rows\":[{\"kind\":8,\"name\":\"Position\",\"url\":\"enums/Position.html\",\"classes\":\"\"},{\"kind\":16,\"name\":\"topLeft\",\"url\":\"enums/Position.html#topLeft\",\"classes\":\"\",\"parent\":\"Position\"},{\"kind\":16,\"name\":\"topCenter\",\"url\":\"enums/Position.html#topCenter\",\"classes\":\"\",\"parent\":\"Position\"},{\"kind\":16,\"name\":\"topRight\",\"url\":\"enums/Position.html#topRight\",\"classes\":\"\",\"parent\":\"Position\"},{\"kind\":16,\"name\":\"bottomLeft\",\"url\":\"enums/Position.html#bottomLeft\",\"classes\":\"\",\"parent\":\"Position\"},{\"kind\":16,\"name\":\"bottomCenter\",\"url\":\"enums/Position.html#bottomCenter\",\"classes\":\"\",\"parent\":\"Position\"},{\"kind\":16,\"name\":\"bottomRight\",\"url\":\"enums/Position.html#bottomRight\",\"classes\":\"\",\"parent\":\"Position\"},{\"kind\":16,\"name\":\"center\",\"url\":\"enums/Position.html#center\",\"classes\":\"\",\"parent\":\"Position\"},{\"kind\":8,\"name\":\"TextBackgroundType\",\"url\":\"enums/TextBackgroundType.html\",\"classes\":\"\"},{\"kind\":16,\"name\":\"stretchX\",\"url\":\"enums/TextBackgroundType.html#stretchX\",\"classes\":\"\",\"parent\":\"TextBackgroundType\"},{\"kind\":16,\"name\":\"stretchY\",\"url\":\"enums/TextBackgroundType.html#stretchY\",\"classes\":\"\",\"parent\":\"TextBackgroundType\"},{\"kind\":16,\"name\":\"none\",\"url\":\"enums/TextBackgroundType.html#none\",\"classes\":\"\",\"parent\":\"TextBackgroundType\"},{\"kind\":8,\"name\":\"ImageFormat\",\"url\":\"enums/ImageFormat.html\",\"classes\":\"\"},{\"kind\":16,\"name\":\"png\",\"url\":\"enums/ImageFormat.html#png\",\"classes\":\"\",\"parent\":\"ImageFormat\"},{\"kind\":16,\"name\":\"jpg\",\"url\":\"enums/ImageFormat.html#jpg\",\"classes\":\"\",\"parent\":\"ImageFormat\"},{\"kind\":16,\"name\":\"base64\",\"url\":\"enums/ImageFormat.html#base64\",\"classes\":\"\",\"parent\":\"ImageFormat\"},{\"kind\":256,\"name\":\"PositionOptions\",\"url\":\"interfaces/PositionOptions.html\",\"classes\":\"\"},{\"kind\":1024,\"name\":\"X\",\"url\":\"interfaces/PositionOptions.html#X\",\"classes\":\"\",\"parent\":\"PositionOptions\"},{\"kind\":1024,\"name\":\"Y\",\"url\":\"interfaces/PositionOptions.html#Y\",\"classes\":\"\",\"parent\":\"PositionOptions\"},{\"kind\":1024,\"name\":\"position\",\"url\":\"interfaces/PositionOptions.html#position\",\"classes\":\"\",\"parent\":\"PositionOptions\"},{\"kind\":256,\"name\":\"TextStyle\",\"url\":\"interfaces/TextStyle.html\",\"classes\":\"\"},{\"kind\":1024,\"name\":\"color\",\"url\":\"interfaces/TextStyle.html#color\",\"classes\":\"\",\"parent\":\"TextStyle\"},{\"kind\":1024,\"name\":\"fontName\",\"url\":\"interfaces/TextStyle.html#fontName\",\"classes\":\"\",\"parent\":\"TextStyle\"},{\"kind\":1024,\"name\":\"fontSize\",\"url\":\"interfaces/TextStyle.html#fontSize\",\"classes\":\"\",\"parent\":\"TextStyle\"},{\"kind\":1024,\"name\":\"shadowStyle\",\"url\":\"interfaces/TextStyle.html#shadowStyle\",\"classes\":\"\",\"parent\":\"TextStyle\"},{\"kind\":1024,\"name\":\"textBackgroundStyle\",\"url\":\"interfaces/TextStyle.html#textBackgroundStyle\",\"classes\":\"\",\"parent\":\"TextStyle\"},{\"kind\":1024,\"name\":\"underline\",\"url\":\"interfaces/TextStyle.html#underline\",\"classes\":\"\",\"parent\":\"TextStyle\"},{\"kind\":1024,\"name\":\"skewX\",\"url\":\"interfaces/TextStyle.html#skewX\",\"classes\":\"\",\"parent\":\"TextStyle\"},{\"kind\":1024,\"name\":\"strikeThrough\",\"url\":\"interfaces/TextStyle.html#strikeThrough\",\"classes\":\"\",\"parent\":\"TextStyle\"},{\"kind\":1024,\"name\":\"textAlign\",\"url\":\"interfaces/TextStyle.html#textAlign\",\"classes\":\"\",\"parent\":\"TextStyle\"},{\"kind\":1024,\"name\":\"italic\",\"url\":\"interfaces/TextStyle.html#italic\",\"classes\":\"\",\"parent\":\"TextStyle\"},{\"kind\":1024,\"name\":\"bold\",\"url\":\"interfaces/TextStyle.html#bold\",\"classes\":\"\",\"parent\":\"TextStyle\"},{\"kind\":1024,\"name\":\"rotate\",\"url\":\"interfaces/TextStyle.html#rotate\",\"classes\":\"\",\"parent\":\"TextStyle\"},{\"kind\":256,\"name\":\"ShadowLayerStyle\",\"url\":\"interfaces/ShadowLayerStyle.html\",\"classes\":\"\"},{\"kind\":1024,\"name\":\"dx\",\"url\":\"interfaces/ShadowLayerStyle.html#dx\",\"classes\":\"\",\"parent\":\"ShadowLayerStyle\"},{\"kind\":1024,\"name\":\"dy\",\"url\":\"interfaces/ShadowLayerStyle.html#dy\",\"classes\":\"\",\"parent\":\"ShadowLayerStyle\"},{\"kind\":1024,\"name\":\"radius\",\"url\":\"interfaces/ShadowLayerStyle.html#radius\",\"classes\":\"\",\"parent\":\"ShadowLayerStyle\"},{\"kind\":1024,\"name\":\"color\",\"url\":\"interfaces/ShadowLayerStyle.html#color\",\"classes\":\"\",\"parent\":\"ShadowLayerStyle\"},{\"kind\":256,\"name\":\"TextBackgroundStyle\",\"url\":\"interfaces/TextBackgroundStyle.html\",\"classes\":\"\"},{\"kind\":1024,\"name\":\"paddingX\",\"url\":\"interfaces/TextBackgroundStyle.html#paddingX\",\"classes\":\"\",\"parent\":\"TextBackgroundStyle\"},{\"kind\":1024,\"name\":\"paddingY\",\"url\":\"interfaces/TextBackgroundStyle.html#paddingY\",\"classes\":\"\",\"parent\":\"TextBackgroundStyle\"},{\"kind\":1024,\"name\":\"type\",\"url\":\"interfaces/TextBackgroundStyle.html#type\",\"classes\":\"\",\"parent\":\"TextBackgroundStyle\"},{\"kind\":1024,\"name\":\"color\",\"url\":\"interfaces/TextBackgroundStyle.html#color\",\"classes\":\"\",\"parent\":\"TextBackgroundStyle\"},{\"kind\":256,\"name\":\"TextOptions\",\"url\":\"interfaces/TextOptions.html\",\"classes\":\"\"},{\"kind\":1024,\"name\":\"text\",\"url\":\"interfaces/TextOptions.html#text\",\"classes\":\"\",\"parent\":\"TextOptions\"},{\"kind\":1024,\"name\":\"positionOptions\",\"url\":\"interfaces/TextOptions.html#positionOptions\",\"classes\":\"\",\"parent\":\"TextOptions\"},{\"kind\":1024,\"name\":\"style\",\"url\":\"interfaces/TextOptions.html#style\",\"classes\":\"\",\"parent\":\"TextOptions\"},{\"kind\":256,\"name\":\"TextMarkOptions\",\"url\":\"interfaces/TextMarkOptions.html\",\"classes\":\"\"},{\"kind\":1024,\"name\":\"backgroundImage\",\"url\":\"interfaces/TextMarkOptions.html#backgroundImage\",\"classes\":\"\",\"parent\":\"TextMarkOptions\"},{\"kind\":1024,\"name\":\"watermarkTexts\",\"url\":\"interfaces/TextMarkOptions.html#watermarkTexts\",\"classes\":\"\",\"parent\":\"TextMarkOptions\"},{\"kind\":1024,\"name\":\"scale\",\"url\":\"interfaces/TextMarkOptions.html#scale\",\"classes\":\"\",\"parent\":\"TextMarkOptions\"},{\"kind\":1024,\"name\":\"quality\",\"url\":\"interfaces/TextMarkOptions.html#quality\",\"classes\":\"\",\"parent\":\"TextMarkOptions\"},{\"kind\":1024,\"name\":\"filename\",\"url\":\"interfaces/TextMarkOptions.html#filename\",\"classes\":\"\",\"parent\":\"TextMarkOptions\"},{\"kind\":1024,\"name\":\"saveFormat\",\"url\":\"interfaces/TextMarkOptions.html#saveFormat\",\"classes\":\"\",\"parent\":\"TextMarkOptions\"},{\"kind\":1024,\"name\":\"maxSize\",\"url\":\"interfaces/TextMarkOptions.html#maxSize\",\"classes\":\"\",\"parent\":\"TextMarkOptions\"},{\"kind\":256,\"name\":\"ImageOptions\",\"url\":\"interfaces/ImageOptions.html\",\"classes\":\"\"},{\"kind\":1024,\"name\":\"src\",\"url\":\"interfaces/ImageOptions.html#src\",\"classes\":\"\",\"parent\":\"ImageOptions\"},{\"kind\":1024,\"name\":\"scale\",\"url\":\"interfaces/ImageOptions.html#scale\",\"classes\":\"\",\"parent\":\"ImageOptions\"},{\"kind\":1024,\"name\":\"rotate\",\"url\":\"interfaces/ImageOptions.html#rotate\",\"classes\":\"\",\"parent\":\"ImageOptions\"},{\"kind\":1024,\"name\":\"alpha\",\"url\":\"interfaces/ImageOptions.html#alpha\",\"classes\":\"\",\"parent\":\"ImageOptions\"},{\"kind\":256,\"name\":\"ImageMarkOptions\",\"url\":\"interfaces/ImageMarkOptions.html\",\"classes\":\"\"},{\"kind\":1024,\"name\":\"backgroundImage\",\"url\":\"interfaces/ImageMarkOptions.html#backgroundImage\",\"classes\":\"\",\"parent\":\"ImageMarkOptions\"},{\"kind\":1024,\"name\":\"watermarkImage\",\"url\":\"interfaces/ImageMarkOptions.html#watermarkImage\",\"classes\":\"\",\"parent\":\"ImageMarkOptions\"},{\"kind\":1024,\"name\":\"watermarkPositions\",\"url\":\"interfaces/ImageMarkOptions.html#watermarkPositions\",\"classes\":\"\",\"parent\":\"ImageMarkOptions\"},{\"kind\":1024,\"name\":\"quality\",\"url\":\"interfaces/ImageMarkOptions.html#quality\",\"classes\":\"\",\"parent\":\"ImageMarkOptions\"},{\"kind\":1024,\"name\":\"filename\",\"url\":\"interfaces/ImageMarkOptions.html#filename\",\"classes\":\"\",\"parent\":\"ImageMarkOptions\"},{\"kind\":1024,\"name\":\"saveFormat\",\"url\":\"interfaces/ImageMarkOptions.html#saveFormat\",\"classes\":\"\",\"parent\":\"ImageMarkOptions\"},{\"kind\":1024,\"name\":\"maxSize\",\"url\":\"interfaces/ImageMarkOptions.html#maxSize\",\"classes\":\"\",\"parent\":\"ImageMarkOptions\"},{\"kind\":128,\"name\":\"Marker\",\"url\":\"classes/Marker.html\",\"classes\":\"\"},{\"kind\":2048,\"name\":\"markText\",\"url\":\"classes/Marker.html#markText\",\"classes\":\"\",\"parent\":\"Marker\"},{\"kind\":2048,\"name\":\"markImage\",\"url\":\"classes/Marker.html#markImage\",\"classes\":\"\",\"parent\":\"Marker\"}],\"index\":{\"version\":\"2.3.9\",\"fields\":[\"name\",\"comment\"],\"fieldVectors\":[[\"name/0\",[0,33.604]],[\"comment/0\",[]],[\"name/1\",[1,38.712]],[\"comment/1\",[]],[\"name/2\",[2,38.712]],[\"comment/2\",[]],[\"name/3\",[3,38.712]],[\"comment/3\",[]],[\"name/4\",[4,38.712]],[\"comment/4\",[]],[\"name/5\",[5,38.712]],[\"comment/5\",[]],[\"name/6\",[6,38.712]],[\"comment/6\",[]],[\"name/7\",[7,38.712]],[\"comment/7\",[]],[\"name/8\",[8,38.712]],[\"comment/8\",[]],[\"name/9\",[9,38.712]],[\"comment/9\",[]],[\"name/10\",[10,38.712]],[\"comment/10\",[]],[\"name/11\",[11,38.712]],[\"comment/11\",[]],[\"name/12\",[12,38.712]],[\"comment/12\",[]],[\"name/13\",[13,38.712]],[\"comment/13\",[]],[\"name/14\",[14,38.712]],[\"comment/14\",[]],[\"name/15\",[15,38.712]],[\"comment/15\",[]],[\"name/16\",[16,33.604]],[\"comment/16\",[]],[\"name/17\",[17,38.712]],[\"comment/17\",[]],[\"name/18\",[18,38.712]],[\"comment/18\",[]],[\"name/19\",[0,33.604]],[\"comment/19\",[]],[\"name/20\",[19,38.712]],[\"comment/20\",[]],[\"name/21\",[20,30.239]],[\"comment/21\",[]],[\"name/22\",[21,38.712]],[\"comment/22\",[]],[\"name/23\",[22,38.712]],[\"comment/23\",[]],[\"name/24\",[23,38.712]],[\"comment/24\",[]],[\"name/25\",[24,33.604]],[\"comment/25\",[]],[\"name/26\",[25,38.712]],[\"comment/26\",[]],[\"name/27\",[26,38.712]],[\"comment/27\",[]],[\"name/28\",[27,38.712]],[\"comment/28\",[]],[\"name/29\",[28,38.712]],[\"comment/29\",[]],[\"name/30\",[29,38.712]],[\"comment/30\",[]],[\"name/31\",[30,38.712]],[\"comment/31\",[]],[\"name/32\",[31,33.604]],[\"comment/32\",[]],[\"name/33\",[32,38.712]],[\"comment/33\",[]],[\"name/34\",[33,38.712]],[\"comment/34\",[]],[\"name/35\",[34,38.712]],[\"comment/35\",[]],[\"name/36\",[35,38.712]],[\"comment/36\",[]],[\"name/37\",[20,30.239]],[\"comment/37\",[]],[\"name/38\",[24,33.604]],[\"comment/38\",[]],[\"name/39\",[36,38.712]],[\"comment/39\",[]],[\"name/40\",[37,38.712]],[\"comment/40\",[]],[\"name/41\",[38,38.712]],[\"comment/41\",[]],[\"name/42\",[20,30.239]],[\"comment/42\",[]],[\"name/43\",[39,38.712]],[\"comment/43\",[]],[\"name/44\",[40,38.712]],[\"comment/44\",[]],[\"name/45\",[16,33.604]],[\"comment/45\",[]],[\"name/46\",[41,38.712]],[\"comment/46\",[]],[\"name/47\",[42,38.712]],[\"comment/47\",[]],[\"name/48\",[43,33.604]],[\"comment/48\",[]],[\"name/49\",[44,38.712]],[\"comment/49\",[]],[\"name/50\",[45,33.604]],[\"comment/50\",[]],[\"name/51\",[46,33.604]],[\"comment/51\",[]],[\"name/52\",[47,33.604]],[\"comment/52\",[]],[\"name/53\",[48,33.604]],[\"comment/53\",[]],[\"name/54\",[49,33.604]],[\"comment/54\",[]],[\"name/55\",[50,38.712]],[\"comment/55\",[]],[\"name/56\",[51,38.712]],[\"comment/56\",[]],[\"name/57\",[45,33.604]],[\"comment/57\",[]],[\"name/58\",[31,33.604]],[\"comment/58\",[]],[\"name/59\",[52,38.712]],[\"comment/59\",[]],[\"name/60\",[53,38.712]],[\"comment/60\",[]],[\"name/61\",[43,33.604]],[\"comment/61\",[]],[\"name/62\",[54,38.712]],[\"comment/62\",[]],[\"name/63\",[55,38.712]],[\"comment/63\",[]],[\"name/64\",[46,33.604]],[\"comment/64\",[]],[\"name/65\",[47,33.604]],[\"comment/65\",[]],[\"name/66\",[48,33.604]],[\"comment/66\",[]],[\"name/67\",[49,33.604]],[\"comment/67\",[]],[\"name/68\",[56,38.712]],[\"comment/68\",[]],[\"name/69\",[57,38.712]],[\"comment/69\",[]],[\"name/70\",[58,38.712]],[\"comment/70\",[]]],\"invertedIndex\":[[\"alpha\",{\"_index\":52,\"name\":{\"59\":{}},\"comment\":{}}],[\"backgroundimage\",{\"_index\":43,\"name\":{\"48\":{},\"61\":{}},\"comment\":{}}],[\"base64\",{\"_index\":15,\"name\":{\"15\":{}},\"comment\":{}}],[\"bold\",{\"_index\":30,\"name\":{\"31\":{}},\"comment\":{}}],[\"bottomcenter\",{\"_index\":5,\"name\":{\"5\":{}},\"comment\":{}}],[\"bottomleft\",{\"_index\":4,\"name\":{\"4\":{}},\"comment\":{}}],[\"bottomright\",{\"_index\":6,\"name\":{\"6\":{}},\"comment\":{}}],[\"center\",{\"_index\":7,\"name\":{\"7\":{}},\"comment\":{}}],[\"color\",{\"_index\":20,\"name\":{\"21\":{},\"37\":{},\"42\":{}},\"comment\":{}}],[\"dx\",{\"_index\":33,\"name\":{\"34\":{}},\"comment\":{}}],[\"dy\",{\"_index\":34,\"name\":{\"35\":{}},\"comment\":{}}],[\"filename\",{\"_index\":47,\"name\":{\"52\":{},\"65\":{}},\"comment\":{}}],[\"fontname\",{\"_index\":21,\"name\":{\"22\":{}},\"comment\":{}}],[\"fontsize\",{\"_index\":22,\"name\":{\"23\":{}},\"comment\":{}}],[\"imageformat\",{\"_index\":12,\"name\":{\"12\":{}},\"comment\":{}}],[\"imagemarkoptions\",{\"_index\":53,\"name\":{\"60\":{}},\"comment\":{}}],[\"imageoptions\",{\"_index\":50,\"name\":{\"55\":{}},\"comment\":{}}],[\"italic\",{\"_index\":29,\"name\":{\"30\":{}},\"comment\":{}}],[\"jpg\",{\"_index\":14,\"name\":{\"14\":{}},\"comment\":{}}],[\"marker\",{\"_index\":56,\"name\":{\"68\":{}},\"comment\":{}}],[\"markimage\",{\"_index\":58,\"name\":{\"70\":{}},\"comment\":{}}],[\"marktext\",{\"_index\":57,\"name\":{\"69\":{}},\"comment\":{}}],[\"maxsize\",{\"_index\":49,\"name\":{\"54\":{},\"67\":{}},\"comment\":{}}],[\"none\",{\"_index\":11,\"name\":{\"11\":{}},\"comment\":{}}],[\"paddingx\",{\"_index\":36,\"name\":{\"39\":{}},\"comment\":{}}],[\"paddingy\",{\"_index\":37,\"name\":{\"40\":{}},\"comment\":{}}],[\"png\",{\"_index\":13,\"name\":{\"13\":{}},\"comment\":{}}],[\"position\",{\"_index\":0,\"name\":{\"0\":{},\"19\":{}},\"comment\":{}}],[\"positionoptions\",{\"_index\":16,\"name\":{\"16\":{},\"45\":{}},\"comment\":{}}],[\"quality\",{\"_index\":46,\"name\":{\"51\":{},\"64\":{}},\"comment\":{}}],[\"radius\",{\"_index\":35,\"name\":{\"36\":{}},\"comment\":{}}],[\"rotate\",{\"_index\":31,\"name\":{\"32\":{},\"58\":{}},\"comment\":{}}],[\"saveformat\",{\"_index\":48,\"name\":{\"53\":{},\"66\":{}},\"comment\":{}}],[\"scale\",{\"_index\":45,\"name\":{\"50\":{},\"57\":{}},\"comment\":{}}],[\"shadowlayerstyle\",{\"_index\":32,\"name\":{\"33\":{}},\"comment\":{}}],[\"shadowstyle\",{\"_index\":23,\"name\":{\"24\":{}},\"comment\":{}}],[\"skewx\",{\"_index\":26,\"name\":{\"27\":{}},\"comment\":{}}],[\"src\",{\"_index\":51,\"name\":{\"56\":{}},\"comment\":{}}],[\"stretchx\",{\"_index\":9,\"name\":{\"9\":{}},\"comment\":{}}],[\"stretchy\",{\"_index\":10,\"name\":{\"10\":{}},\"comment\":{}}],[\"strikethrough\",{\"_index\":27,\"name\":{\"28\":{}},\"comment\":{}}],[\"style\",{\"_index\":41,\"name\":{\"46\":{}},\"comment\":{}}],[\"text\",{\"_index\":40,\"name\":{\"44\":{}},\"comment\":{}}],[\"textalign\",{\"_index\":28,\"name\":{\"29\":{}},\"comment\":{}}],[\"textbackgroundstyle\",{\"_index\":24,\"name\":{\"25\":{},\"38\":{}},\"comment\":{}}],[\"textbackgroundtype\",{\"_index\":8,\"name\":{\"8\":{}},\"comment\":{}}],[\"textmarkoptions\",{\"_index\":42,\"name\":{\"47\":{}},\"comment\":{}}],[\"textoptions\",{\"_index\":39,\"name\":{\"43\":{}},\"comment\":{}}],[\"textstyle\",{\"_index\":19,\"name\":{\"20\":{}},\"comment\":{}}],[\"topcenter\",{\"_index\":2,\"name\":{\"2\":{}},\"comment\":{}}],[\"topleft\",{\"_index\":1,\"name\":{\"1\":{}},\"comment\":{}}],[\"topright\",{\"_index\":3,\"name\":{\"3\":{}},\"comment\":{}}],[\"type\",{\"_index\":38,\"name\":{\"41\":{}},\"comment\":{}}],[\"underline\",{\"_index\":25,\"name\":{\"26\":{}},\"comment\":{}}],[\"watermarkimage\",{\"_index\":54,\"name\":{\"62\":{}},\"comment\":{}}],[\"watermarkpositions\",{\"_index\":55,\"name\":{\"63\":{}},\"comment\":{}}],[\"watermarktexts\",{\"_index\":44,\"name\":{\"49\":{}},\"comment\":{}}],[\"x\",{\"_index\":17,\"name\":{\"17\":{}},\"comment\":{}}],[\"y\",{\"_index\":18,\"name\":{\"18\":{}},\"comment\":{}}]],\"pipeline\":[]}}");
