+++ docs/localization.json/ImageFormat.json
@@ -1,0 +1,15 @@
+{
+    "ImageFormat": {
+        "comment": {
+            "blockTags": [],
+            "modifierTags": {},
+            "summary": [
+                {
+                    "kind": "text",
+                    "text": "ImageFormat enum for save image format"
+                }
+            ]
+        },
+        "members": {}
+    }
+}
