+++ docs/localization.json/ImageMarkOptions.json
@@ -1,0 +1,101 @@
+{
+    "ImageMarkOptions": {
+        "comment": {
+            "blockTags": [],
+            "modifierTags": {},
+            "summary": [
+                {
+                    "kind": "text",
+                    "text": "Options for image watermark"
+                }
+            ]
+        },
+        "properties": {
+            "quality": {
+                "comment": {
+                    "blockTags": [
+                        {
+                            "tag": "@defaultValue",
+                            "content": [
+                                {
+                                    "kind": "code",
+                                    "text": "`1`"
+                                },
+                                {
+                                    "kind": "text",
+                                    "text": "\nimage quality 0 - 1"
+                                }
+                            ]
+                        }
+                    ],
+                    "modifierTags": {},
+                    "summary": []
+                }
+            },
+            "saveFormat": {
+                "comment": {
+                    "blockTags": [
+                        {
+                            "tag": "@defaultValue",
+                            "content": [
+                                {
+                                    "kind": "code",
+                                    "text": "`jpg`"
+                                },
+                                {
+                                    "kind": "text",
+                                    "text": "\nsave image format see"
+                                }
+                            ]
+                        },
+                        {
+                            "tag": "@ImageFormat",
+                            "content": [
+                                {
+                                    "kind": "text",
+                                    "text": "eg. 'png'"
+                                }
+                            ]
+                        }
+                    ],
+                    "modifierTags": {},
+                    "summary": []
+                }
+            },
+            "maxSize": {
+                "comment": {
+                    "blockTags": [
+                        {
+                            "tag": "@defaultValue",
+                            "content": [
+                                {
+                                    "kind": "text",
+                                    "text": "2048\nandroid only\n**need RN version >= 0.60.0**,  fresco "
+                                },
+                                {
+                                    "kind": "code",
+                                    "text": "`MaxBitmapSize`"
+                                },
+                                {
+                                    "kind": "text",
+                                    "text": " ["
+                                },
+                                {
+                                    "kind": "code",
+                                    "text": "`ImagePipelineConfig.Builder.experiment().setMaxBitmapSize()`"
+                                },
+                                {
+                                    "kind": "text",
+                                    "text": "](https://github.com/facebook/fresco/blob/08ca5f40cc0b60b4db16d15e45552cafeae39ccb/imagepipeline/src/main/java/com/facebook/imagepipeline/core/ImagePipelineExperiments.java#L282), see [#49](https://github.com/JimmyDaddy/react-native-image-marker/issues/49#issuecomment-535303838)"
+                                }
+                            ]
+                        }
+                    ],
+                    "modifierTags": {},
+                    "summary": []
+                }
+            }
+        },
+        "methods": {}
+    }
+}
