+++ docs/localization.json/ImageOptions.json
@@ -1,0 +1,80 @@
+{
+    "ImageOptions": {
+        "comment": {
+            "blockTags": [],
+            "modifierTags": {},
+            "summary": [
+                {
+                    "kind": "text",
+                    "text": "Image options for image watermark"
+                }
+            ]
+        },
+        "properties": {
+            "scale": {
+                "comment": {
+                    "blockTags": [
+                        {
+                            "tag": "@defaultValue",
+                            "content": [
+                                {
+                                    "kind": "code",
+                                    "text": "`1`"
+                                },
+                                {
+                                    "kind": "text",
+                                    "text": "\nscale image 0 - 1"
+                                }
+                            ]
+                        }
+                    ],
+                    "modifierTags": {},
+                    "summary": []
+                }
+            },
+            "rotate": {
+                "comment": {
+                    "blockTags": [
+                        {
+                            "tag": "@defaultValue",
+                            "content": [
+                                {
+                                    "kind": "code",
+                                    "text": "`0`"
+                                },
+                                {
+                                    "kind": "text",
+                                    "text": "\nrotate image with degree 0 - 360"
+                                }
+                            ]
+                        }
+                    ],
+                    "modifierTags": {},
+                    "summary": []
+                }
+            },
+            "alpha": {
+                "comment": {
+                    "blockTags": [
+                        {
+                            "tag": "@defaultValue",
+                            "content": [
+                                {
+                                    "kind": "code",
+                                    "text": "`1`"
+                                },
+                                {
+                                    "kind": "text",
+                                    "text": "\nimage alpha 0 - 1"
+                                }
+                            ]
+                        }
+                    ],
+                    "modifierTags": {},
+                    "summary": []
+                }
+            }
+        },
+        "methods": {}
+    }
+}
