+++ docs/localization.json/TextBackgroundStyle.json
@@ -1,0 +1,16 @@
+{
+    "TextBackgroundStyle": {
+        "comment": {
+            "blockTags": [],
+            "modifierTags": {},
+            "summary": [
+                {
+                    "kind": "text",
+                    "text": "background style for text watermark\n> thanks [@onka13](https://github.com/onka13) for [#38](https://github.com/JimmyDaddy/react-native-image-marker/pull/38)"
+                }
+            ]
+        },
+        "properties": {},
+        "methods": {}
+    }
+}
