+++ docs/localization.json/globalFunctions.json
@@ -1,0 +1,1 @@
+{}
